did some cleanup on the components
